requirements.txt
discord.py==1.6.0
python-dotenv==0.15.0
pydash==4.9.2
boto3==1.17.24
nltk==3.5
pandas==1.2.3
requests==2.25.1
unidecode==1.2.0
git+https://github.com/Rapptz/discord-ext-menus